+++ internal/playground-realtime-api/src/task/index.ts
@@ -1,31 +1,33 @@
-import { Task } from '@signalwire/realtime-api'
-
-const client = new Task.Client({
-  host: process.env.HOST || 'relay.swire.io',
-  project: process.env.PROJECT as string,
-  token: process.env.TOKEN as string,
-  contexts: ['office'],
-  debug: {
-    logWsTraffic: true,
-  },
-})
-
-client.on('task.received', (payload) => {
-  console.log('Task Received', payload)
-})
-
-setTimeout(async () => {
-  console.log('Sending to the client..')
-  await Task.send({
+import { SignalWire } from '@signalwire/realtime-api'
+;(async () => {
+  const client = await SignalWire({
     host: process.env.HOST || 'relay.swire.io',
     project: process.env.PROJECT as string,
     token: process.env.TOKEN as string,
-    context: 'office',
+  })
+
+  const removeOfficeListeners = await client.task.listen({
+    topics: ['office', 'home'],
+    onTaskReceived: (payload) => {
+      console.log('Task received under the "office" or "home" context', payload)
+    },
+  })
+
+  console.log('Sending a message to office..')
+  await client.task.send({
+    topic: 'office',
     message: { yo: ['bro', 1, true] },
   })
 
+  console.log('Sending a message to home..')
+  await client.task.send({
+    topic: 'home',
+    message: { yo: ['bro', 2, true] },
+  })
+
   setTimeout(async () => {
+    await removeOfficeListeners()
     console.log('Disconnect the client..')
     client.disconnect()
   }, 2000)
-}, 2000)
+})()
