--- packages/realtime-api/src/signalwire/index.ts
@@ -1,1 +1,0 @@
-export { SignalWire } from './SignalWire'
