+++ packages/realtime-api/src/task/Task.ts
@@ -1,58 +1,83 @@
+import { request } from 'node:https'
 import {
-  DisconnectableClientContract,
-  BaseComponentOptions,
-  ClientContextContract,
-  ApplyEventListeners,
+  EventEmitter,
+  TaskInboundEvent,
+  TaskReceivedEventName,
 } from '@signalwire/core'
-import { connect } from '@signalwire/core'
-import type { TaskClientApiEvents } from '../types'
-import { RealtimeClient } from '../client/index'
+import { SWClient } from '../SWClient'
 import { taskWorker } from './workers/taskWorker'
+import { BaseListenOptions, BaseNamespace } from '../BaseNamespace'
 
-export interface Task
-  extends DisconnectableClientContract<Task, TaskClientApiEvents>,
-    ClientContextContract {
-  /** @internal */
-  _session: RealtimeClient
-  /**
-   * Disconnects this client. The client will stop receiving events and you will
-   * need to create a new instance if you want to use it again.
-   *
-   * @example
-   *
-   * ```js
-   * client.disconnect()
-   * ```
-   */
-  disconnect(): void
+const PATH = '/api/relay/rest/tasks'
+const HOST = 'relay.signalwire.com'
+
+interface TaskListenOptions extends BaseListenOptions {
+  onTaskReceived?: (payload: TaskInboundEvent['message']) => unknown
 }
 
-/** @internal */
-class TaskAPI extends ApplyEventListeners<TaskClientApiEvents> {
-  constructor(options: BaseComponentOptions<TaskClientApiEvents>) {
+type TaskListenersKeys = keyof Omit<TaskListenOptions, 'topics'>
+
+export class Task extends BaseNamespace<TaskListenOptions> {
+  protected emitter = new EventEmitter()
+  protected _eventMap: Record<TaskListenersKeys, TaskReceivedEventName> = {
+    onTaskReceived: 'task.received',
+  }
+
+  constructor(options: SWClient) {
     super(options)
 
-    this.runWorker('taskWorker', {
+    this._client.runWorker('taskWorker', {
       worker: taskWorker,
+      initialState: {
+        emitter: this.emitter,
+      },
     })
-    this._attachListeners('')
+  }
+
+  send({
+    topic,
+    message,
+  }: {
+    topic: string
+    message: Record<string, unknown>
+  }) {
+    const { userOptions } = this._sw
+    if (!userOptions.project || !userOptions.token) {
+      throw new Error('Invalid options: project and token are required!')
+    }
+    return new Promise<void>((resolve, reject) => {
+      try {
+        const Authorization = `Basic ${Buffer.from(
+          `${userOptions.project}:${userOptions.token}`
+        ).toString('base64')}`
+
+        const data = JSON.stringify({ context: topic, message })
+        const options = {
+          host: userOptions.host ?? HOST,
+          port: 443,
+          method: 'POST',
+          path: PATH,
+          headers: {
+            Authorization,
+            'Content-Type': 'application/json',
+            'Content-Length': data.length,
+          },
+        }
+        const req = request(options, ({ statusCode }) => {
+          statusCode === 204 ? resolve() : reject()
+        })
+
+        req.on('error', reject)
+
+        req.write(data)
+        req.end()
+      } catch (error) {
+        reject(error)
+      }
+    })
   }
 }
 
-/** @internal */
-export const createTaskObject = (
-  params: BaseComponentOptions<TaskClientApiEvents>
-): Task => {
-  const task = connect<TaskClientApiEvents, TaskAPI, Task>({
-    store: params.store,
-    Component: TaskAPI,
-  })(params)
-
-  return task
-}
-
-export * from './TaskClient'
-export * from './send'
 export type { TaskReceivedEventName } from '@signalwire/core'
 export type {
   TaskClientApiEvents,
